fix&del: allow decimal angle & delete grade col in layer

+++ app/Http/Requests/StoreLayerRequest.php
@@ -28,7 +28,7 @@ class StoreLayerRequest extends FormRequest
             ],
             'thickness' => ['required', 'numeric', 'min:0'],
             'width'     => ['required', 'numeric', 'min:0'],
-            'angle'     => ['required', 'integer', 'between:-180,180'],
+            'angle'     => ['required', 'numeric', 'between:-180,180'],
         ];
     }
 }
